+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* New enums and allocators for Isochronous endpoints
 
-...
+### Changed
+* `EndpointType` enum now has fields for isochronous synchronization and usage.
 
 ## [0.2.9] - 2022-08-02